gtasks_cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gtasks-cli"
version = "0.1.3"
authors = [
    {name = "Google Tasks CLI Team", email = "example@example.com"},
]
description = "A powerful command-line interface for managing Google Tasks with enhanced features and functionality which is secure and easy to use with advanced features."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "click>=8.1.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "rich>=13.0.0",
    "tabulate>=0.9.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "pyyaml>=6.0",
    "appdirs>=1.4.4",
    "cryptography>=41.0.0",
    "requests>=2.31.0",
    "matplotlib>=3.7.0",
    "prompt-toolkit>=3.0.0",
    "colorama>=0.4.6",
    "litellm>=1.42.0",
    "mcp>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/example/gtasks-cli"
"Bug Reports" = "https://github.com/example/gtasks-cli/issues"
"Source" = "https://github.com/example/gtasks-cli"

[project.scripts]
gtasks = "gtasks_cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black",
    "flake8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gtasks_cli"]